+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # UPCOMING
 * Show only valid actions for the current site.
+* Use GM_setClipboard to set the clipboard.
 
 # v0.5
 * Support Amazon CN.